% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_beta.R
\name{betadisper}
\alias{betadisper}
\title{...}
\usage{
betadisper(dm, physeq, variable)
}
\value{
.
}
\description{
...
}
\note{
.

.

.
}
\examples{

library(phyloseq)
data(esophagus)
phyloseq_compute_bdiv(esophagus, 100) -> dist





}
\author{
Florentin Constancias
}
