% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_beta.R
\name{calc_pairwise_permanovas_strata}
\alias{calc_pairwise_permanovas_strata}
\title{...}
\usage{
calc_pairwise_permanovas_strata(
  dm,
  metadata_map,
  compare_header,
  n_perm,
  strat
)
}
\value{
.
}
\description{
...
}
\note{
.

.

.
}
\examples{

library(phyloseq)
data(esophagus)
phyloseq_compute_bdiv(esophagus, 100) -> dist






}
\author{
Florentin Constancias
}
