% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_alpha.R
\name{correlate_alpha}
\alias{correlate_alpha}
\title{...}
\usage{
correlate_alpha(alphas_all, colnums_to_plot, colour, shape, method)
}
\value{
.
}
\description{
...
}
\note{
.remove ASV assigned to Chloroplast and Mitochondria

.

.
}
\examples{

library(phyloseq)
data("enterotype")




#"spearman"
}
\author{
Florentin Constancias
}
