% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_beta.R
\name{phyloseq_adonis_strata_perm}
\alias{phyloseq_adonis_strata_perm}
\title{...}
\usage{
phyloseq_adonis_strata_perm(
  dm,
  physeq = physeq,
  formula = paste0(variables, collapse = " + "),
  nrep = nrep,
  strata = strata
)
}
\value{
.
}
\description{
...
}
\note{
.

.

.
}
\examples{

library(phyloseq)
data(esophagus)
phyloseq_compute_bdiv(esophagus, 100) -> dist





}
\author{
Florentin Constancias
}
