% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_alpha.R
\name{phyloseq_alphas}
\alias{phyloseq_alphas}
\title{...}
\usage{
phyloseq_alphas(physeq, phylo)
}
\value{
.
}
\description{
...
}
\note{
.

.

.
}
\examples{
library(phyloseq)
library(tidyverse)
data("esophagus")

phyloseq_alphas(esophagus) -> results

}
\author{
Florentin Constancias
}
