% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_heatmap.R
\name{phyloseq_ampvis_heatmap}
\alias{phyloseq_ampvis_heatmap}
\title{...}
\usage{
phyloseq_ampvis_heatmap(
  physeq,
  transform,
  group_by,
  facet_by,
  tax_aggregate,
  tax_add,
  ntax
)
}
\value{
.
}
\description{
...
}
\note{
.

.

.
}
\examples{

library(phyloseq)
data("GlobalPatterns")

phyloseq_ampvis_heatmap(GlobalPatterns,"percent", "Primer","SampleType",  "Species", "Phylum", 10) -> out






}
\author{
Florentin Constancias
}
