% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_normalisation.R
\name{phyloseq_check_lib_size}
\alias{phyloseq_check_lib_size}
\title{...}
\usage{
phyloseq_check_lib_size(
  physeq,
  data_color,
  data_facet,
  nreads_display,
  first_n
)
}
\value{
.
}
\description{
...
}
\note{
.

.

.
}
\examples{

library(phyloseq)
data("enterotype")

phyloseq_check_lib_size(enterotype,"SeqTech","Project", 1000, 10) -> out


}
\author{
Florentin Constancias
}
