% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_beta.R
\name{phyloseq_compute_bdiv}
\alias{phyloseq_compute_bdiv}
\title{...}
\usage{
phyloseq_compute_bdiv(phylo_tmp, norm, phylo, seed)
}
\value{
.
}
\description{
...
}
\note{
.

.

.
}
\examples{

library(phyloseq)
data(esophagus)
phyloseq_compute_bdiv(esophagus, 100) -> dist






}
\author{
Florentin Constancias
}
