% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_taxa_tests.R
\name{phyloseq_correlate_taxa}
\alias{phyloseq_correlate_taxa}
\title{...}
\usage{
phyloseq_correlate_taxa(
  ps_tmp,
  grouping_column,
  adjustment = 3,
  cor_variables,
  method
)
}
\value{
.
}
\description{
...
}
\note{
. correlates ASV/Taxa with metadata

.

.
}
\examples{

library(phyloseq)
data("enterotype")




#"spearman"

}
\author{
Florentin Constancias
}
