% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_normalisation.R
\name{phyloseq_density_normalize}
\alias{phyloseq_density_normalize}
\title{...}
\usage{
phyloseq_density_normalize(
  physeq = physeq,
  qPCR = physeq \%>\% sample_data() \%>\% data.frame() \%>\% rownames_to_column("tmp"),
  sample_idx = "tmp",
  value_idx = "ReadNorm"
)
}
\value{
.
}
\description{
...
}
\note{
.

.

.
}
\examples{

library(phyloseq)
data("GlobalPatterns")
sample_data(GlobalPatterns)$norm = sample(5000000:1000000, size=nsamples(GlobalPatterns), replace=TRUE)
phyloseq_density_normalize(GlobalPatterns, "norm") -> out
HTSSIP::OTU_qPCR_trans(GlobalPatterns, GlobalPatterns \%>\% sample_data() \%>\% data.frame() , "X.SampleID", value_idx = "norm") -> out2
otu_table(GlobalPatterns)["540305","TRRsed3"] #2305
sample_data(GlobalPatterns)["TRRsed3",] #2664122
sample_sums(GlobalPatterns)["TRRsed3"] #279704
(2305/279704)  * 2664122
otu_table(out)["540305","TRRsed3"] #
otu_table(out2)["540305","TRRsed3"] #
}
\author{
Florentin Constancias
}
