% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_normalisation.R
\name{phyloseq_filter_samples}
\alias{phyloseq_filter_samples}
\title{Identify and Filter ASV from phyloseq object / dataframe with samples as rows.}
\usage{
phyloseq_filter_samples(physeq, thrs)
}
\arguments{
\item{physeq}{phyloseq object / dataframe with samples as rows}

\item{thrs}{threshold in \%}
}
\value{
list of 2 data frames / phyloseq objects of rare - common ASV
}
\description{
Identify and Filter ASV from phyloseq object / dataframe with samples as rows.
}
\note{
For each sample, determine cumulated sum of percentage of sorted ASV.

Then, define Common ASV for wich cumulated sum of percentag is less or equal to percentage threshold.

As ASV can be rare / common in different samples, ASV are defined as common if they are at least common in on samples. The rares ASV are the remaining ones.
}
\examples{
library(phyloseq)
library(tidyverse)
data("esophagus")

# Data table with taxa samples as rows
esophagus \%>\%
otu_table() \%>\%
as.data.frame() -> data

phyloseq_filter_samples(data, 50) -> results

}
\author{
Florentin Constancias
}
