% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_normalisation.R
\name{phyloseq_get_strains}
\alias{phyloseq_get_strains}
\title{...}
\usage{
phyloseq_get_strains(physeq)
}
\value{
.
}
\description{
...
}
\note{
.

.

.
}
\examples{

library(phyloseq)
data("enterotype")

phyloseq_check_lib_size(enterotype,"SeqTech","Project", 1000, 10) -> out


}
\author{
Florentin Constancias
}
