% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_beta.R
\name{phyloseq_plot_PCoA_3d}
\alias{phyloseq_plot_PCoA_3d}
\title{...}
\usage{
phyloseq_plot_PCoA_3d(ps_rare, dlist, m = "PCoA", seed, color, shape)
}
\value{
.
}
\description{
...
}
\note{
.

.

.
}
\examples{

library(phyloseq)
data(esophagus)
phyloseq_compute_bdiv(esophagus, 100) -> dist







}
\author{
Florentin Constancias
}
