% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_normalisation.R
\name{phyloseq_rarefaction_curves}
\alias{phyloseq_rarefaction_curves}
\title{...}
\usage{
phyloseq_rarefaction_curves(physeq, stepsize, color_data, facet_data)
}
\value{
.
}
\description{
...
}
\note{
.

.

.
}
\examples{

library(phyloseq)
data("enterotype")

phyloseq_rarefaction_curves(GlobalPatterns, 10000,  "Primer", "SampleType") -> out


}
\author{
Florentin Constancias
}
