% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_normalisation.R
\name{phyloseq_rarefied_unrarefied_richness}
\alias{phyloseq_rarefied_unrarefied_richness}
\title{...}
\usage{
phyloseq_rarefied_unrarefied_richness(
  physeq,
  sample_size,
  seed,
  color_data,
  fill_data,
  shape_data
)
}
\value{
.
}
\description{
...
}
\note{
.

.

.
}
\examples{

library(phyloseq)
data("enterotype")

phyloseq_rarefied_unrerefied_richness(GlobalPatterns, 123,  "Primer", "Primer", "SampleType") -> out






}
\author{
Florentin Constancias
}
