% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_normalisation.R
\name{phyloseq_remove_chloro_mitho}
\alias{phyloseq_remove_chloro_mitho}
\title{...}
\usage{
phyloseq_remove_chloro_mitho(physeq)
}
\value{
.
}
\description{
...
}
\note{
.remove ASV assigned to Chloroplast and Mitochondria

.

.
}
\examples{

library(phyloseq)
data("enterotype")




}
\author{
Florentin Constancias
}
