% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_taxa_tests.R
\name{phyloseq_run_compare_means}
\alias{phyloseq_run_compare_means}
\title{...}
\usage{
phyloseq_run_compare_means(
  tmp = tmp,
  group = group,
  comp = comp,
  prev = prev,
  varcoef = varcoef
)
}
\value{
.
}
\description{
...
}
\note{
.

.

.
}
\examples{

library(phyloseq)
data(esophagus)
phyloseq_compute_bdiv(esophagus, 100) -> dist






}
\author{
Florentin Constancias
}
