% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_beta.R
\name{physeq_pairwise_permanovas}
\alias{physeq_pairwise_permanovas}
\title{...}
\usage{
physeq_pairwise_permanovas(dm, physeq, compare_header, n_perm, strat)
}
\value{
.
}
\description{
...
}
\note{
.

.

.
}
\examples{

library(phyloseq)
data(esophagus)
phyloseq_compute_bdiv(esophagus, 100) -> dist



}
\author{
Florentin Constancias
}
