% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_alpha.R
\name{plot_alphas}
\alias{plot_alphas}
\title{...}
\usage{
plot_alphas(
  alphas,
  measure,
  x_group,
  colour_group,
  fill_group,
  shape_group,
  facet_group,
  test_group,
  test_group_2
)
}
\value{
.
}
\description{
...
}
\note{
.remove ASV assigned to Chloroplast and Mitochondria

.

.
}
\examples{

}
\author{
Florentin Constancias
}
